requirements.txt
# common
django
daphne
channels

# point_field_app | single_table_app | dynamic_table_app
psycopg

# mongo_app
mongoengine

# tests
pytest
websockets
pytest-asyncio

# automation and reports
colorama
pandas
matplotlib
streamlit
python-dotenv

